removing the failing test for #10
@hijonathan removing this test because it's not clear
that this should be the expected behavior. While it definitely
should not return NaNundefined, I'm not sure what the function
should return when the minimal sane string representation of a
number in units that we handle (k, M, B) is longer than the
'charWdith' specified by the user. In this case, I'm not sure
what `intword(100, 2)` should return since `0.1k` has three
characters before the unit.

Should we just return 100 here?